Fix the operation of q_sort()_v1
Identify the issue as the lack of judgment on whether it is a single-element list,
which prevents the execution of q_sort() recursively.

Modify the implementation method for descending order,
as the original program only implements ascending order.
If it is judged as descending, simply reverse it using q_reverse().